Ported crane button, crane camera and pipe puzzle to proximity Interactable framework
