+++ intertwined/Assets/Scripts/Interactable/ConstructionSite/CraneButton.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 
 namespace Interactable.ConstructionSite
 {
-    public class CraneButton : MonoBehaviour
+    public class CraneButton : Interactable
     {
         private Crane _crane;
         protected void Start()
@@ -11,20 +11,26 @@ namespace Interactable.ConstructionSite
             _crane = GetComponentInParent<Crane>();
         }
 
-        protected void OnTriggerEnter(Collider other)
-        {
-            if (CollidingObjectCanInteract(other)) _crane.NumInTrigger++;
-        }
+        // protected void OnTriggerEnter(Collider other)
+        // {
+        //     if (CollidingObjectCanInteract(other)) _crane.NumInTrigger++;
+        // }
+        //
+        // protected void OnTriggerExit(Collider other)
+        // {
+        //     if (CollidingObjectCanInteract(other)) _crane.NumInTrigger--;
+        // }
+        //
+        // private bool CollidingObjectCanInteract(Component other)
+        // {
+        //     return (other.CompareTag("Boy") || other.CompareTag("BoySubObjects"))
+        //            || (other.CompareTag("Dog") || other.CompareTag("DogSubObjects"));
+        // }
 
-        protected void OnTriggerExit(Collider other)
+        protected override void ProximityInteraction(Character.Character interacter, bool enteredTrigger)
         {
-            if (CollidingObjectCanInteract(other)) _crane.NumInTrigger--;
-        }
-        
-        private bool CollidingObjectCanInteract(Component other)
-        {
-            return (other.CompareTag("Boy") || other.CompareTag("BoySubObjects"))
-                   || (other.CompareTag("Dog") || other.CompareTag("DogSubObjects"));
+            if (enteredTrigger) _crane.NumInTrigger++;
+            else _crane.NumInTrigger--;
         }
     }
 }
